services/campaign/requirements.txt
Flask==0.12
flask-restful==0.3.5
py2neo
python-dotenv